@@ -1,37 +1,46 @@
 {
-  "name": "uglybox2",
-  "version": "0.1.0",
+  "name": "vue-project",
+  "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "nodemon --exec 'vue-cli-service serve'",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "bundle": "node bin/bundle.js"
+    "dev": "vite",
+    "api": "node src/server/server.cjs",
+    "serve": "yarn api & yarn dev",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "@vue-hero-icons/solid": "^1.7.2",
-    "archiver": "^5.3.0",
-    "autoprefixer": "^9",
-    "axios": "^0.21.1",
-    "core-js": "^3.6.5",
-    "dotenv-flow": "^3.2.0",
-    "nodemon": "^2.0.7",
-    "postcss": "^7",
-    "postcss-loader": "^4.0.3",
-    "retry-axios": "^2.4.0",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0"
+    "@heroicons/vue": "^2.1.3",
+    "axios": "^1.6.8",
+    "dotenv-flow": "^4.1.0",
+    "express": "^4.19.2",
+    "pinia": "^2.1.7",
+    "vue": "^3.4.21",
+    "vue-router": "^4.3.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "babel-eslint": "^10.1.0",
-    "babel-preset-vue-app": "^2.0.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "@rushstack/eslint-patch": "^1.3.3",
+    "@tsconfig/node20": "^20.1.2",
+    "@types/node": "^20.11.28",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "@vue/eslint-config-prettier": "^8.0.0",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "@vue/tsconfig": "^0.5.1",
+    "autoprefixer": "^10.4.19",
+    "eslint": "^8.49.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "npm-run-all2": "^6.1.2",
+    "postcss": "^8.4.38",
+    "prettier": "^3.0.3",
+    "tailwindcss": "^3.4.3",
+    "typescript": "~5.4.0",
+    "vite": "^5.1.6",
+    "vue-tsc": "^2.0.6"
   }
 }
