@@ -1,18 +1,19 @@
 import Camera from '../../lib/Camera'
+import { defineStore } from 'pinia'
 
-var AssistToolsStore = {
-  state: { value: '-', opts: [], attribute: 'assitool_exposure' },
+export const useAssistToolsStore = defineStore('assistTools', {
+  state: () => ({ value: '-', opts: [], attribute: 'assitool_exposure', }),
 
-  initialize() {
-    this.get_exposure()
+  actions: {
+    initialize() {
+      this.get_exposure()
+    },
+
+    get_exposure() {
+      Camera.ctrl_get(this.attribute, (response) => {
+        this.value = response.value
+        this.opts = response.opts
+      })
+    },
   },
-
-  get_exposure() {
-    Camera.ctrl_get(this.state.attribute, (response) => {
-      this.state.value = response.value
-      this.state.opts = response.opts
-    })
-  },
-}
-
-export default AssistToolsStore
+})
