@@ -1,18 +1,19 @@
 import Camera from '../../lib/Camera'
+import { defineStore } from 'pinia'
 
-var ResolutionStore = {
-  state: { value: '-', opts: [], attribute: 'resolution' },
+export const useResolutionStore = defineStore('resolution', {
+  state: () => ({ value: '-', opts: [], attribute: 'resolution' }),
 
-  initialize() {
-    this.get_resolution()
+  actions: {
+    initialize() {
+      this.get_resolution()
+    },
+
+    get_resolution() {
+      Camera.ctrl_get(this.attribute, (response) => {
+        this.value = response.value
+        this.opts = response.opts
+      })
+    },
   },
-
-  get_resolution() {
-    Camera.ctrl_get(this.state.attribute, (response) => {
-      this.state.value = response.value
-      this.state.opts = response.opts
-    })
-  },
-}
-
-export default ResolutionStore
+})
